// cli/src/core/api/types.ts
export interface PaginatedResponse<T> {
  totalPages: number;
  totalElements: number;
  data: T[];
}